Fix how we get the `intl` provider.
In production mode only (and possibly depending on Volto and VLT version and which addons you have), I get this error when trying to a add a separator or button block:

```
[React Intl] Could not find required `intl` object. <IntlProvider> needs to exist in the component ancestry.
```

So the strange thing is that I can add a quote block and most other blocks but two blocks not.
And when I remove the quote block from the addons, all is well.

Partially I may have caused this myself last year by this PR:
https://github.com/kitconcept/volto-quote-block/pull/28
That fixed https://github.com/kitconcept/volto-quote-block/issues/27 for me.
But it may have not been the best way.

So we try again.  But even with the changes I did so far, it does not work for me.